Move pdf viewer reducer
