+++ client/app/reader/CommentLayer.jsx
@@ -195,7 +195,7 @@ CommentLayer.propTypes = {
 };
 
 const mapStateToProps = (state, ownProps) => ({
-  ...state.readerReducer.ui.pdf,
+  //...state.readerReducer.ui.pdf,
   ..._.pick(state.annotationLayer, 'placingAnnotationIconPageCoords'),
   comments: makeGetAnnotationsByDocumentId(state)(ownProps.documentId),
   isPlacingAnnotation: state.annotationLayer.isPlacingAnnotation,
