+++ client/app/reader/PdfPage.jsx
@@ -349,7 +349,7 @@ const mapStateToProps = (state, props) => {
     searchText: searchText(state, props),
     currentMatchIndex: getCurrentMatchIndex(state, props),
     matchesPerPage: getMatchesPerPageInFile(state, props),
-    searchBarHidden: state.readerReducer.ui.pdf.hideSearchBar
+    searchBarHidden: state.pdfViewer.hideSearchBar
   };
 };
