+++ client/app/reader/PdfSidebar.jsx
@@ -29,6 +29,7 @@ import { commentColumns, commentInstructions, documentsColumns,
 import classNames from 'classnames';
 import { makeGetAnnotationsByDocumentId } from './selectors';
 import { CATEGORIES } from './analytics';
+import { COMMENT_ACCORDION_KEY } from '../reader/PdfViewer/actionTypes';
 
 const COMMENT_SCROLL_FROM_THE_TOP = 50;
 
@@ -171,7 +172,7 @@ export class PdfSidebar extends React.Component {
             <SideBarIssueTags
               doc={this.props.doc} />
           </AccordionSection>
-          <AccordionSection title={Constants.COMMENT_ACCORDION_KEY} id="comments-header">
+          <AccordionSection title={COMMENT_ACCORDION_KEY} id="comments-header">
             <SideBarComments
               comments={comments}
             />
@@ -256,10 +257,10 @@ const mapStateToProps = (state, ownProps) => {
     ..._.pick(state.annotationLayer, 'placedButUnsavedAnnotation', 'selectedAnnotationId'),
     comments: makeGetAnnotationsByDocumentId(state)(ownProps.doc.id),
     scrollToSidebarComment: state.readerReducer.ui.pdf.scrollToSidebarComment,
-    hidePdfSidebar: state.readerReducer.ui.pdf.hidePdfSidebar,
-    error: state.readerReducer.ui.pdfSidebar.error,
+    error: state.pdfViewer.pdfSideBarError,
     appeal: state.readerReducer.loadedAppeal,
-    ..._.pick(state.readerReducer, 'documents', 'openedAccordionSections')
+    openedAccordionSections: state.pdfViewer.openedAccordionSections,
+    hidePdfSidebar: state.pdfViewer.hidePdfSidebar
   };
 };
 const mapDispatchToProps = (dispatch) => ({
