+++ client/app/reader/PdfUI.jsx
@@ -289,6 +289,7 @@ const mapStateToProps = (state, props) => {
     docListIsFiltered: docListIsFiltered(state),
     loadError: state.readerReducer.documentErrors[props.doc.content_url],
     isPlacingAnnotation: state.annotationLayer.isPlacingAnnotation,
+    ..._.pick(state.pdfViewer, 'hidePdfSidebar'),
     ...state.readerReducer.ui.pdf,
     numPages
   };
