+++ client/app/reader/commonActions.js
@@ -3,7 +3,7 @@ import { getUpdatedFilteredResults } from './searchFilters';
 import {
   HIDE_ERROR_MESSAGE,
   SHOW_ERROR_MESSAGE
-} from './constants';
+} from './PdfViewer/actionTypes';
 
 import { UPDATE_FILTERED_RESULTS } from './DocumentList/actionTypes';
 import { ASSIGN_DOCUMENTS } from './Documents/actionTypes';
